+++ src/lib.rs
@@ -36,7 +36,7 @@ pub mod prelude {
     pub use crate::cooldown::{Cooldown, Cooldowns};
     pub use crate::input_map::InputMap;
     pub use crate::input_mocking::MockInput;
-    pub use crate::user_input::UserInput;
+    pub use crate::user_input::{Modifier, UserInput};
 
     pub use crate::plugin::InputManagerPlugin;
     pub use crate::plugin::ToggleActions;